Bring all DbConnectionOptions files into the respective projects, move all instances to ConnectionString namespace
