+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/Common/ConnectionString/DbConnectionOptions.netcore.cs
@@ -5,9 +5,8 @@
 using System;
 using System.IO;
 using System.Text;
-using Microsoft.Data.Common.ConnectionString;
 
-namespace Microsoft.Data.Common
+namespace Microsoft.Data.Common.ConnectionString
 {
     internal partial class DbConnectionOptions
     {
